Splitting the hell out of everything!
Please test all of this for bugz
