+++ balance_cmd.py
@@ -1,0 +1,4 @@
+import economy
+
+def return_command_response(message):
+    return '{}, your balance is: {}'.format(message.author, economy.get_user_balance(message.author))
