+++ economy.py
@@ -1,0 +1,11 @@
+user_balance_dict = {}
+
+def get_user_balance(user):
+    if str(user) in user_balance_dict:
+        return user_balance_dict[str(user)]
+    else:
+        user_balance_dict[str(user)] = 0
+        return get_user_balance(user)
+
+def set_user_balance(user, balance):
+    user_balance_dict[str(user)] = balance;
