+++ hello_cmd.py
@@ -1,0 +1,2 @@
+def return_command_response(message):
+    return 'Hello {}'.format(message.author)
